+++ src/normalize.rs
@@ -1,6 +1,6 @@
 use super::consts::{ALL_MIN, HUE_MAX, PERCENT_MAX, RATIO_MAX, RGB_UNIT_MAX};
 
-fn normalize(val: f32, max: f32) -> f32 {
+fn normalize(val: f64, max: f64) -> f64 {
   if val < ALL_MIN {
     return ALL_MIN;
   }
@@ -10,28 +10,28 @@ fn normalize(val: f32, max: f32) -> f32 {
   val
 }
 
-pub fn normalize_percent(val: f32) -> f32 {
+pub fn normalize_percent(val: f64) -> f64 {
   normalize(val, PERCENT_MAX)
 }
 
-pub fn normalize_hue(h: f32) -> f32 {
+pub fn normalize_hue(h: f64) -> f64 {
   let h = normalize(h, HUE_MAX);
-  if (h - HUE_MAX).abs() < std::f32::EPSILON {
+  if (h - HUE_MAX).abs() < std::f64::EPSILON {
     0.0
   } else {
     h
   }
 }
 
-pub fn normalize_rgb_unit(val: f32) -> f32 {
+pub fn normalize_rgb_unit(val: f64) -> f64 {
   normalize(val, RGB_UNIT_MAX)
 }
 
-pub fn normalize_ratio(val: f32) -> f32 {
+pub fn normalize_ratio(val: f64) -> f64 {
   normalize(val, RATIO_MAX)
 }
 
-pub fn bound(r: f32, entire: f32) -> f32 {
+pub fn bound(r: f64, entire: f64) -> f64 {
   let mut n = r;
   loop {
     let less = n < ALL_MIN;
@@ -47,14 +47,14 @@ pub fn bound(r: f32, entire: f32) -> f32 {
   }
 }
 
-pub fn bound_ratio(r: f32) -> f32 {
+pub fn bound_ratio(r: f64) -> f64 {
   bound(r, RATIO_MAX)
 }
 
-pub fn bound_hue(h: f32) -> f32 {
+pub fn bound_hue(h: f64) -> f64 {
   bound(h, HUE_MAX)
 }
 
-pub fn round_ratio(r: f32) -> f32 {
+pub fn round_ratio(r: f64) -> f64 {
   (r * 100.0).round() / 100.0
 }
